Move jobs into their own files Fix NAT problem with redirects Add appmetrics, measurement metrics
