--- hybrasyl/Jobs.cs
@@ -1,399 +1,0 @@
-﻿/*
- * This file is part of Project Hybrasyl.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the Affero General Public License as published by
- * the Free Software Foundation, version 3.
- *
- * This program is distributed in the hope that it will be useful, but
- * without ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
- * or FITNESS FOR A PARTICULAR PURPOSE. See the Affero General Public License
- * for more details.
- *
- * You should have received a copy of the Affero General Public License along
- * with this program. If not, see <http://www.gnu.org/licenses/>.
- *
- * (C) 2020 ERISCO, LLC 
- *
- * For contributors and individual authors please refer to CONTRIBUTORS.MD.
- * 
- */
-
-using Hybrasyl.Enums;
-using Hybrasyl.Objects;
-using System;
-using System.Linq;
-using System.Timers;
-
-namespace Hybrasyl
-{
-    namespace Jobs
-    {
-        // Jobs (each one is a class) to be scheduled at startup by Hybrasy's timers go here.
-        // NOTE: ONLY JOB CLASSES GO HERE.
-        //
-        // Each class needs an Interval (which represents how often it should run, in seconds)
-        // and a void Execute() which will do the work. If you want logging (and you should)
-        // you need the following as well:
-        //
-        //        public static readonly ILog Logger =
-        //       LogManager.GetLogger(
-        //       System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
-        //
-        // Please note that a static logger is the ONLY way to guarantee threadsafe logging.
-        //
-        // Jobs run in their own thread, so there are some important considerations.
-        //
-        // 1) Never modify game state or call game logic directly. That's what message passing is for!
-        //    Reading state is OK, unless you're using it for something critical (player state)
-        //    in which case consistency (obviously) isn't guaranteed.
-        //    Example: a job that, say, occasionally reported the number of logged in players to
-        //    somewhere else via an external API call: probably fine. A job that checked to see if
-        //    a player had a certain item and then took action based on that - BAD. The only exception
-        //    to this is if the object implements some kind of locking (see Mailbox for an example).
-        //
-        // 2) You can send packets to clients from a job (since doing so is intended to be
-        //    thread safe). An instance of how to use this is in the AutoSnoreJob, where we can
-        //    send snore packets to idle clients without being a bother to anything else. Since this
-        //    is effectively a threadsafe operation that doesn't change game state or logic - this
-        //    is fine.
-        //
-        // The best way to use a job is to write a packet handler for a control message, and have
-        // the job submit a control message. CheckpointerJob is a great example of this.
-        //
-
-        public static class CheckpointerJob
-        {
-            public static readonly int Interval = 300;
-
-            public static void Execute(object obj, ElapsedEventArgs args)
-            {
-                try
-                {
-                    GameLog.Debug("Job starting");
-                    foreach (var client in GlobalConnectionManifest.WorldClients)
-                    {
-                        // Insert a "save client" message onto the queue for each client.
-                        // We do this rather than sending a "checkpoint" message so we don't
-                        // randomly have a packet occupying shitloads of CPU time blocking
-                        // everything else.
-
-                        World.ControlMessageQueue.Add(new HybrasylControlMessage(ControlOpcodes.SaveUser, client.Key));
-                    }
-                    GameLog.Debug("Job complete");
-                }
-                catch (Exception e)
-                {
-                    Game.ReportException(e);
-                    GameLog.Error("Exception occured in job:", e);
-                }
-            }
-        }
-
-        public static class ShutdownJob
-        {
-            public static readonly int Interval = 10;
-
-            public static void Execute(object obj, ElapsedEventArgs args)
-            {
-                if (Game.ShutdownTimeRemaining == -1) return; // No shutdown yet requested
-                else if (Game.ShutdownTimeRemaining == 0)
-                {
-                    // Shutdown has arrived
-                    World.ControlMessageQueue.Add(new HybrasylControlMessage(ControlOpcodes.ShutdownServer, "job", 0));
-                    return;
-                }
-                else if (Game.ShutdownTimeRemaining <= 300)
-                {
-                    // Send message every minute
-                    if (Game.ShutdownTimeRemaining % 60 == 0)
-                    {
-                        foreach (var user in Game.World.ActiveUsers)
-                            user.SendSystemMessage($"Chaos will be rising up in {Game.ShutdownTimeRemaining / 60} minute(s).");
-                    }
-                }
-                else if (Game.ShutdownTimeRemaining > 600)
-                {
-                    // Send message every five minutes
-                    if (Game.ShutdownTimeRemaining % 300 == 0)
-                    {
-                        foreach (var user in Game.World.ActiveUsers)
-                            user.SendSystemMessage($"Chaos will be rising up in {Game.ShutdownTimeRemaining % 60} minutes.");
-                    }
-                }
-                Game.ShutdownTimeRemaining -= Interval;
-            }
-        }
-
-        public static class MailboxCleanupJob
-        {
-            // Clean up mailboxes once an hour
-            public static readonly int Interval = 3600;
-
-            public static void Execute(object obj, ElapsedEventArgs args)
-            {
-                try
-                {
-                    GameLog.Debug("Job starting");
-
-                    var now = DateTime.Now.Ticks;
-                    foreach (var mailbox in Game.World.WorldData.Values<Mailbox>().Where(mb => mb.Full))
-                    {
-                        try
-                        {
-                            mailbox.Lock();
-                            mailbox.Cleanup();
-                            mailbox.Unlock();
-                        }
-                        catch (MessageStoreLocked e)
-                        {
-                            Game.ReportException(e);
-                            GameLog.ErrorFormat("{0}: mailbox locked during cleanup...?", mailbox.Name);
-                        }
-                    }
-                    foreach (var board in Game.World.WorldData.Values<Board>().Where(mb => mb.Full))
-                    {
-                        try
-                        {
-                            board.Lock();
-                            board.Cleanup();
-                            board.Unlock();
-                        }
-                        catch (MessageStoreLocked e)
-                        {
-                            Game.ReportException(e);
-                            GameLog.ErrorFormat("{0}: board locked during cleanup...?", board.Name);
-                        }
-                    }
-                    GameLog.Debug("Job complete");
-                }
-                catch (Exception e)
-                {
-                    Game.ReportException(e);
-                    GameLog.Error("Exception occured in job:", e);
-                }
-            }
-        }
-
-        public static class IdleDetectionJob
-        {
-
-            public static readonly int Interval = 60;
-
-            public static void Execute(object obj, ElapsedEventArgs args)
-            {
-                try
-                {
-                    GameLog.Debug("Job starting");
-
-                    var now = DateTime.Now.Ticks;
-                    foreach (var client in GlobalConnectionManifest.WorldClients.Values)
-                    {
-                        client.CheckIdle();
-                    }
-                    GameLog.Debug("Job complete");
-                }
-                catch (Exception e)
-                {
-                    Game.ReportException(e);
-                    GameLog.Error("Exception occured in job:", e);
-                }
-            }
-        }
-
-        public static class ByteHeartbeatJob
-        {
-
-            public static readonly int Interval = Constants.BYTE_HEARTBEAT_INTERVAL;
-
-            public static void Execute(object obj, ElapsedEventArgs args)
-            {
-                try
-                {
-                    GameLog.Debug("Job starting");
-
-                    var rnd = new Random();
-                    foreach (var client in GlobalConnectionManifest.WorldClients.Values)
-                    {
-                        // Send the 0x3B heartbeat to logged in clients
-                        client.SendByteHeartbeat();
-                    }
-                    GameLog.Debug("Job complete");
-                }
-                catch (Exception e)
-                {
-                    Game.ReportException(e);
-                    GameLog.Error("Exception occured in job:", e);
-                }
-            }
-        }
-
-        public static class TickHeartbeatJob
-        {
-
-            public static readonly int Interval = Constants.TICK_HEARTBEAT_INTERVAL;
-
-            public static void Execute(object obj, ElapsedEventArgs args)
-            {
-                try
-                {
-                    GameLog.Debug("Job starting");
-
-                    var rnd = new Random();
-                    foreach (var client in GlobalConnectionManifest.WorldClients.Values)
-                    {
-                        client.SendTickHeartbeat();
-                    }
-                    GameLog.Debug("Job complete");
-                }
-                catch (Exception e)
-                {
-                    Game.ReportException(e);
-                    GameLog.Error("Exception occured in job:", e);
-                }
-            }
-        }
-
-        /// <summary>
-        /// This job reaps connections that haven't responded to either tick or byte based heartbeats
-        /// in REAP_HEARTBEAT_TIME.
-        /// </summary>
-        public static class HeartbeatReaperJob
-        {
-
-            public static readonly int Interval = Constants.REAP_HEARTBEAT_INTERVAL;
-
-            public static void Execute(object obj, ElapsedEventArgs args)
-            {
-                GameLog.Debug("Job starting");
-                try
-                {
-                    foreach (var connection in GlobalConnectionManifest.WorldClients)
-                    {
-                        var client = connection.Value;
-                        var connectionId = connection.Key;
-                        User user;
-                        if (Game.World.WorldData.TryGetValueByIndex(connectionId, out user))
-                        {
-                            if (client.IsHeartbeatExpired())
-                            {
-                                GameLog.InfoFormat("{0} (connection id {1}: heartbeat expired, disconnecting",
-                                    user.Name, connectionId);
-                                GlobalConnectionManifest.DeregisterClient(client);
-                                World.ControlMessageQueue.Add(new HybrasylControlMessage(ControlOpcodes.CleanupUser, 
-                                    CleanupType.ByConnectionId, connectionId));
-                            }
-                        }
-                    }
-                }
-                catch (Exception e)
-                {
-                    Game.ReportException(e);
-                    GameLog.ErrorFormat("Exception occurred in job:", e);
-                }
-            }
-        }
-
-        public static class RegenerationJob
-        {
-            public static readonly int Interval = 25;
-
-            public static void Execute(object obj, ElapsedEventArgs args)
-            {
-                foreach (var connId in GlobalConnectionManifest.WorldClients.Keys)
-                {
-                    World.ControlMessageQueue.Add(new HybrasylControlMessage(ControlOpcodes.RegenUser, connId));
-                }
-            }
-        }
-
-        public static class StatusTickJob
-        {
-            public static readonly int Interval = 1;
-
-            public static void Execute(object obj, ElapsedEventArgs args)
-            {
-                GameLog.Debug("Status tick job starting");
-                foreach (var connectionId in GlobalConnectionManifest.WorldClients.Keys)
-                {
-                    User user;
-                    if (Game.World.WorldData.TryGetValueByIndex(connectionId, out user))
-                    {
-                        if (user.Statuses.Count > 0 && user.Condition.Alive)
-                            World.ControlMessageQueue.Add(new HybrasylControlMessage(ControlOpcodes.StatusTick, user.Id));
-                    }
-                }
-                foreach (var wobj in Game.World.ActiveStatuses)
-                {
-                    if (wobj is Creature creature && creature.Condition.Alive)
-                        World.ControlMessageQueue.Add(new HybrasylControlMessage(ControlOpcodes.StatusTick, wobj.Id));
-                }
-                GameLog.Debug("Status tick job ending");
-
-            }
-        }
-
-        public static class AutoSnoreJob
-        {
-            public static readonly int Interval = 10;
-
-            public static void Execute(object obj, ElapsedEventArgs args)
-            {
-                GameLog.Debug("Job starting");
-                try
-                {
-                    // FIXME: make this more efficient / don't break our own conventions
-                    foreach (var connection in GlobalConnectionManifest.WorldClients)
-                    {
-                        var client = connection.Value;
-                        var connectionId = connection.Key;
-
-                        if (client.IsIdle())
-                        {
-                            User user;
-                            if (Game.World.WorldData.TryGetValueByIndex(connectionId, out user))
-                            {
-                                user.Motion(16, 120); // send snore effect
-                            }
-                            else
-                            {
-                                GameLog.WarningFormat(
-                                    "Connection id {0} marked as idle but no corresponding user found...?",
-                                    connectionId);
-                            }
-                        }
-                    }
-
-                    GameLog.Debug("Job complete");
-                }
-                catch (Exception e)
-                {
-                    Game.ReportException(e);
-                    GameLog.Error("Exception occured in job:", e);
-                }
-            }
-        }
-
-        public static class MerchantInventoryRefreshJob
-        {
-            public static readonly int Interval = 60;
-
-            public static void Execute(object obj, ElapsedEventArgs args)
-            {
-                GameLog.Debug("MerchantInventoryRefresh Job starting");
-                try
-                {
-                    foreach (Merchant merchant in Game.World.Objects.Values.Where(x => x is Merchant))
-                    {
-                        merchant.RestockInventory();
-                    }
-                }
-                catch (Exception e)
-                {
-                    Game.ReportException(e);
-                    GameLog.Error("Exception occured in MerchantInventoryRefresh job:", e);
-                }
-            }
-        }
-  }
-}
