+++ hybrasyl/World.cs
@@ -45,7 +45,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Timers;
 using System.Xml.Schema;
-
+using System.Diagnostics;
 
 namespace Hybrasyl
 {
@@ -4382,8 +4382,9 @@ namespace Hybrasyl
 
         private void QueueConsumer()
         {
+            var stopwatch = new Stopwatch();
             while (!MessageQueue.IsCompleted)
-            {
+            {                
                 if (StopToken.IsCancellationRequested)
                     return;
                 // Process messages.
@@ -4392,6 +4393,7 @@ namespace Hybrasyl
                 try
                 {
                     message = MessageQueue.Take();
+                    stopwatch.Start();
                 }
                 catch (InvalidOperationException e)
                 {
@@ -4405,6 +4407,7 @@ namespace Hybrasyl
                 {                   
                     var clientMessage = (HybrasylClientMessage)message;
                     var handler = PacketHandlers[clientMessage.Packet.Opcode];
+                    Game.MetricsStore.Measure.Meter.Mark(HybrasylMetricsRegistry.MessageMeter, $"0x{clientMessage.Packet.Opcode}");
                     try
                     {
                         if (TryGetActiveUserById(clientMessage.ConnectionId, out user))
@@ -4462,10 +4465,17 @@ namespace Hybrasyl
                             // Last but not least, invoke the handler
 
                             handler.Invoke(user, clientMessage.Packet);
+                            stopwatch.Stop();
+                            Game.MetricsStore.Measure.Histogram.Update(HybrasylMetricsRegistry.ServiceTime, 
+                                stopwatch.ElapsedMilliseconds, $"0x{clientMessage.Packet.Opcode}");
+
                         }
                         else if (clientMessage.Packet.Opcode == 0x10) // Handle special case of join world
                         {
                             PacketHandlers[0x10].Invoke(clientMessage.ConnectionId, clientMessage.Packet);
+                            stopwatch.Stop();
+                            Game.MetricsStore.Measure.Histogram.Update(HybrasylMetricsRegistry.ServiceTime,
+                                stopwatch.ElapsedMilliseconds, $"0x{clientMessage.Packet.Opcode}");
                         }
                         else
                         {
@@ -4479,6 +4489,10 @@ namespace Hybrasyl
                     catch (Exception e)
                     {
                         Game.ReportException(e);
+                        stopwatch.Stop();
+                        Game.MetricsStore.Measure.Meter.Mark(HybrasylMetricsRegistry.ExceptionMeter, 
+                            $"0x{clientMessage.Packet.Opcode}");
+
                         GameLog.Error(e, "{Opcode}: Unhandled exception encountered in packet handler!", clientMessage.Packet.Opcode);
                     }
                 }
@@ -4489,33 +4503,44 @@ namespace Hybrasyl
 
         public void ControlQueueConsumer()
         {
+            var stopwatch = new Stopwatch();
             while (!ControlMessageQueue.IsCompleted)
             {
                 if (StopToken.IsCancellationRequested)
                     return;
                 // Process messages.
                 HybrasylMessage message;
+                var startTime = DateTime.Now;
                 try
                 {
                     message = ControlMessageQueue.Take();
+                    stopwatch.Start();
                 }
                 catch (InvalidOperationException e)
                 {
                     Game.ReportException(e);
+                    stopwatch.Stop();
                     GameLog.Error("QUEUE CONSUMER: EXCEPTION RAISED: {exception}", e);
                     continue;
                 }
 
-                if (message is HybrasylControlMessage)
+                if (message is HybrasylControlMessage hcm)
                 {
+                    Game.MetricsStore.Measure.Meter.Mark(HybrasylMetricsRegistry.MessageMeter, $"cm_{hcm.Opcode}");
+
                     try
                     {
-                        var controlMessage = (HybrasylControlMessage)message;
-                        ControlMessageHandlers[controlMessage.Opcode].Invoke(controlMessage);                       
+                        ControlMessageHandlers[hcm.Opcode].Invoke(hcm);
+                        stopwatch.Stop();
+                        Game.MetricsStore.Measure.Histogram.Update(HybrasylMetricsRegistry.ServiceTime,
+                            stopwatch.ElapsedMilliseconds, $"cm_{hcm.Opcode}");
                     }
                     catch (Exception e)
                     {
                         Game.ReportException(e);
+                        stopwatch.Stop();
+                        Game.MetricsStore.Measure.Meter.Mark(HybrasylMetricsRegistry.ExceptionMeter,
+                            $"cm_{hcm.Opcode}");
                         GameLog.Error("Exception encountered in control message handler: {exception}", e);
                     }
                 }
